[instr] dynamic caller-saved registers adjustment
Previously, the register for return value is preserved on stack
for an inserted callee. However, this will ignore the return value
of the inserted callee.

This commit adjust caller-saved registers based on whether the inserted function has
a return value or not. If it does not, then we should not save the
return value register on the stack.