[instr] add AT direction and INCLUDE action.
AT allows one to insert at an instruction such that if a label is at the instruction, the label will be at the new instruction.
INCLUDE allows one to compile and link in a C file, obj file, or a library. INCLUDE takes in an optional command to compile C or ASM.